+++ examples/xor.rs
@@ -28,7 +28,7 @@ fn main() -> NNResult<()> {
 
     println!("Predictions:\n");
 
-    let predictions: Array1<f64> = train_data
+    let predictions: Array1<f32> = train_data
         .rows()
         .into_iter()
         .map(|input| {
